Update FileCache.cs
Added an option to enable hashing of keys, defaulting to not.

GetPath now includes the extension instead of providing file name only. Load and other methods and variables have this change reflected.

Fixed an issue where directory path would change when another instance of the class is created (variable was previously made static).